# openapi-spec.json
{
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/ip-library": {
      "get": {
        "operationId": "IpLibraryController_get",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "page number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_size",
            "required": true,
            "in": "query",
            "description": "page size",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "tag of the ip library",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "search keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "category of the ip library",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "required": false,
            "in": "query",
            "description": "genre of the ip library, you can pass multiple genres by comma",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "sort order, default sort field is created_at",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_field",
            "required": false,
            "in": "query",
            "description": "sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "required": false,
            "in": "query",
            "description": "owner of the ip library",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "launched_to_giggle",
            "required": false,
            "in": "query",
            "description": "is launched to giggle",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          },
          {
            "name": "is_top",
            "required": false,
            "in": "query",
            "description": "is top",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_public",
            "required": false,
            "in": "query",
            "description": "is public",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "children_levels",
            "required": false,
            "in": "query",
            "description": "children levels",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "app-id",
            "required": false,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpLibraryListDto"
                }
              }
            }
          }
        },
        "summary": "Get list of ip libraries",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/my": {
      "get": {
        "operationId": "IpLibraryController_getMy",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "page number",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_size",
            "required": true,
            "in": "query",
            "description": "page size",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "tag of the ip library",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "search keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "category of the ip library",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "genre",
            "required": false,
            "in": "query",
            "description": "genre of the ip library, you can pass multiple genres by comma",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_by",
            "required": false,
            "in": "query",
            "description": "sort order, default sort field is created_at",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_field",
            "required": false,
            "in": "query",
            "description": "sort field",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "required": false,
            "in": "query",
            "description": "owner of the ip library",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "launched_to_giggle",
            "required": false,
            "in": "query",
            "description": "is launched to giggle",
            "schema": {
              "enum": [
                "true",
                "false"
              ],
              "type": "string"
            }
          },
          {
            "name": "is_top",
            "required": false,
            "in": "query",
            "description": "is top",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "is_public",
            "required": false,
            "in": "query",
            "description": "is public",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "children_levels",
            "required": false,
            "in": "query",
            "description": "children levels",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpLibraryListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my ip libraries",
        "tags": [
          "IP Library"
        ]
      }
    },
...
    "/api/v1/ip-library/like": {
      "post": {
        "operationId": "IpLibraryController_likeIp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LikeIpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Like an ip library",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/unlike": {
      "post": {
        "operationId": "IpLibraryController_unlikeIp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnlikeIpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Unlike an ip library",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/available-parent-ips": {
      "get": {
        "operationId": "IpLibraryController_getAvailableParentIps",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableParentIpsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get available parent ips",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/my/{id}": {
      "get": {
        "operationId": "IpLibraryController_getMyDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpLibraryDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get detail of ip library",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/set-visibility": {
      "post": {
        "operationId": "IpLibraryController_setIpVisibility",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVisibilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpLibraryDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set visibility of ip library",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/untokenize": {
      "post": {
        "operationId": "IpLibraryController_untokenize",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UntokenizeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpLibraryDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Untokenize an ip library",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/create-ip": {
      "post": {
        "description": "\nReturns SSE stream with progress updates and final result, \nsse event:\n\n**data structure:**\n`\n{\n  event: string,\n  data: {\n    message: string\n  } | number\n}\n`\n\n**event list:**\n- ip.data_validating\n\nthis event indicate the data is validating\n- ip.asset_processing\n\nthis event indicate the asset is processing\n- ip.video_uploading\n\nthis event indicate the video is uploading, at current step, data in `data` is the progress of video uploading\n- ip.ip_library_creating\n\nthis event indicate the ip library is creating\n- ip.push_ip_to_chain\n\nthis event indicate the ip is pushing to chain\n- ip.share_to_giggle\n  \nthis event indicate the ip is sharing to giggle, this event only exists when `share_to_giggle` is true in request body\n- asset.uploading\n  \nthis event indicate the asset of creating meme is uploading, at current step, data in `data` is the progress of asset uploading, this event only exists when `share_to_giggle` is true in request body\n- meme.creating\n\nthis event indicate the meme is creating, this event only exists when `share_to_giggle` is true in request body\n- meme.created\n\nthis event indicate the meme is created, this event only exists when `share_to_giggle` is true in request body\n- ip.update_token_data_on_chain\n\nthis event indicate the ip is updating token data which just created to chain, this event only exists when `share_to_giggle` is true in request body\n- ip.payment_processing\n\nthis event indicate the ip is processing payment, this event only exists when `share_to_giggle` is true in request body\n- ip.payment_confirmed\n\nthis event indicate the ip is confirmed payment, this event only exists when `share_to_giggle` is true in request body\n- ip.payment_refunded\n\nthis event indicate the ip is refunded payment, this event only exists when `share_to_giggle` is true in request body\n- ip.created\n\nthis event indicate the ip is created, and the data in `data` is the detail of ip\n\n- ip.warning\n\nthis event indicate the ip is created, but some warning occurs, this may ip token is created failed or ip token is registered failed, the data in `data` is the warning message\n\n**error event:**\n\nif error occurs, the event will be `error` and the data in `data` is the error message, subscriber will be completed.\ndata structure:\n\n`\nevent: error\nid: 2\ndata: some error message\n`\n\n",
        "operationId": "IpLibraryController_createIp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "event": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create ip library",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/update-ip": {
      "post": {
        "description": "\nUpdate an existing ip, do not allow to update if ip is on chain or token info is already created.\nReturns SSE stream with progress updates and final result, \nsse event:\n\n**data structure:**\n`\n{\n  event: string,\n  data: {\n    message: string\n  } | number\n}\n`\n\n**event list:**\n- ip.data_validating\n\nthis event indicate the data is validating\n\n- ip.on_chain_updating\n\nthis event indicate the ip is updating on chain.\n\n- ip.updated\n\nthis event indicate the ip is updated, and the data in `data` is the detail of ip\n\n- ip.warning\n\nthis event indicate the ip is updated, but some warning occurs, this may ip token is created failed or ip token is registered failed, the data in `data` is the warning message\n\n**error event:**\n\nif error occurs, the event will be `error` and the data in `data` is the error message, subscriber will be completed.\ndata structure:\n\n`\nevent: error\nid: 2\ndata: some error message\n`\n\n",
        "operationId": "IpLibraryController_editIp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditIpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SSEMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update ip",
        "tags": [
          "IP Library"
        ]
      }
    },
...
    "/api/v1/ip-library/{id}": {
      "get": {
        "operationId": "IpLibraryController_getDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpLibraryDetailDto"
                }
              }
            }
          }
        },
        "summary": "Get detail of ip library",
        "tags": [
          "IP Library"
        ]
      }
    },
    "/api/v1/ip-library/register-token": {
      "post": {
        "operationId": "IpLibraryController_registerToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register token for an ip",
        "tags": [
          "IP Library"
        ]
      }
    },
...
  },
  "info": {
    "title": "3 Body Labs API Reference",
    "description": "This is the API reference for 3 Body Labs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "JWT Authorization"
      }
    },
    "schemas": {
...
      "IpLibraryListDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "ip libraries",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IpSummaryWithChildDto"
            }
          },
          "count": {
            "type": "number",
            "description": "total count of ip libraries"
          }
        },
        "required": [
          "data",
          "count"
        ]
      },
      "LikeIpDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id of the ip library"
          }
        },
        "required": [
          "id"
        ]
      },
      "UnlikeIpDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id of the ip library"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "AvailableParentIpsDto": {
        "type": "object",
        "properties": {
          "owned": {
            "description": "owned ip libraries",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AvailableParentIpDto"
            }
          },
          "purchased": {
            "description": "purchased ip libraries",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchasedIpDto"
            }
          }
        },
        "required": [
          "owned",
          "purchased"
        ]
      },
      "SetVisibilityDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id of the ip library"
          },
          "is_public": {
            "type": "boolean",
            "description": "visibility of the ip library"
          }
        },
        "required": [
          "id",
          "is_public"
        ]
      },
      "UntokenizeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "id of the ip library"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  },
...
}